Add files via upload
Differential Extruder uses two closed "loop" timing belts that engage a timing pulley in the print head from two opposite sides. When both belts pull in the same direction, the pulley can't spin and belts pull the print head along the axis of the movement. When belts pull in opposite directions to each other, they spin the timing pulley (which spins the extruder wheel in the print head, as shown in the attached video), whilst the print head itself does not move. Please print the attached STL, obtain the required parts and follow the photos for the assembly. Please feel free to modify anything you want. This is just an example that should familiarize you with the new concept and I'm sure that you'll find solutions that will be much more appropriate for the printer you'd like to build or modify. The most elegant way to implement the Differential Extruder will require re-programming of the firmware, because steppers will have to receive commands that make them spin together in order to deliver linear movement of the print head and/or the spinning of the extruder wheel. 
I will shortly upload a translation device that will enable you to retrofit the Differential Extruder onto i3-style 3d printers without firmware re-programming. It is a bit bulky, but it works. You can see it action in the attached video.
I have made this particular incarnation of the Differential Extruder for an i3-style printer because that's what I use and know. However, there should be no problem to implement this same concept on other kinematics such as CoreXY and H-bot. 
The "Allparts" photo shows the parts I used for the print head carriage: 
2 x 20T GT2 timing pulleys (short ones with grub screws between the teeth) 
2 x 5mm nylon-style bearings (feel free to use ball bearings instead) 
4 x 16T smooth idlers, 3mm shaft dia 
4 x M3 stand-offs, 9mm tall 
4 x M3 bolts, 32mm long 
4 x M3 nuts 
3d printed print head case, as per the uploaded STL  

You'll also need at least: 
2 x 20T GT2 timing pulleys, to engage the belts 
2 x 20T GT2 idlers, to tension the belts